refactor: modularize index.ts from 4995 to 557 lines
Split monolithic index.ts into 14 focused modules:
- src/runtime-state.ts: shared runtime counters and rate limit maps
- src/middleware/auth.ts: Bearer token authentication
- src/middleware/rate-limit.ts: IP and distributed rate limiting
- src/helpers/crypto.ts: sha256, stableStringify
- src/helpers/format.ts: formatOutput() consolidating 4x DRY violation
- src/helpers/response.ts: CSP constants, error responses
- src/handlers/convert.ts: core convertUrl() split into 6 sub-functions
- src/handlers/stream.ts: SSE streaming
- src/handlers/health.ts: /api/health endpoint
- src/handlers/batch.ts: /api/batch endpoint
- src/handlers/extract.ts: /api/extract endpoint
- src/handlers/deepcrawl.ts: /api/deepcrawl endpoint
- src/handlers/jobs.ts: /api/jobs + JobCoordinator Durable Object
- src/handlers/og-image.ts: OG image generation

Key improvements:
- convertUrl() broken into orchestrator + 6 sub-functions for readability
- formatOutput() eliminates 4x repeated switch(format) pattern
- Prepares codebase for open-source release (Apache-2.0)

Test results: 527/530 pass (3 pre-existing failures unchanged)
